:art: use double instead of int for pos of coins and obstacle

+++ src/shared/include/Coin.hpp
@@ -1,10 +1,10 @@
 #pragma once
 
 typedef struct coinsPos_s {
-    int x_pos;
-    int y_pos;
+    double x_pos;
+    double y_pos;
     size_t coinId;
 
-    coinsPos_s(int x = 0, int y = 0, size_t id = 0)
+    coinsPos_s(double x = 0.0, double y = 0.0, size_t id = 0)
         : x_pos(x), y_pos(y), coinId(id) {}
 } coinsPos_t;
